# internal/interface/api/spec/openapi.yaml
openapi: 3.1.0
# TODO: split into multiple files: parameters.yaml, models.yaml, etc
info:
  title: Energy Consumption API
  description: |-
    Report the energy consumption of a set of electricity meters.

  contact:
    email: daniel.fernandez3@u.icesi.edu.co
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http:localhost:8081
tags:
  - name: consumption
    description: Operations related to energy consumption.
paths:
  /consumption:
    get:
      tags:
        - consumption
      summary: Get consumption report.
      description: Returns the energy consumption for each meter for the period between startDate and endDate.
      operationId: getConsumption
      parameters: 
        - name: meter_id
          in: query
          description: Meter ids to filter by.
          explode: true # ?meter_id=1,meter_id=2
          required: true
          schema:
            type: array
            items:
              type: integer
              format: uint8
        - name: start_date
          in: query
          description: "Start date (format: YYYY-MM-DD)."
          example: 2017-07-21
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: "End date (format: YYYY-MM-DD)."
          example: 2026-01-19
          required: true
          schema:
            type: string
            format: date
        - name: period
          in: query
          description: "Period (any of [monthly, weekly, daily])."
          example: monthly
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/PeriodicConsumption'
        '400':
          description: Invalid query params
          content:
            application/json:
              schema:
                $ref: "./models.yaml#/components/schemas/Error"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "./models.yaml#/components/schemas/Error"
      security:
        - bearer_auth: []

components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT